chore: hide visitor implementations
